bin/bin/requirements.txt
opencv-python
numpy
scipy
mediapipe
flask
pillow
requests
pythonosc